Renaming and reorganizing classes.
