+++ SpeedRunApp.UI/src/components/charts/LeaderboardGridWorldRecordChart.vue
@@ -18,7 +18,7 @@
     FusionCharts.addDep(MSLine, InverseMSLine, CandyTheme);
 
     export default {
-        name: "GameSpeedRunGridWorldRecordChart",
+        name: "LeaderboardGridWorldRecordChart",
         props: {  
             tabledata: Array,
             title: String,
