+++ SpeedRunApp.UI/src/components/game/WorldRecordGrid.vue
@@ -40,7 +40,7 @@
     import { scrollBehaviourDragImageTranslateOverride } from "mobile-drag-drop/scroll-behaviour";
 
     export default {
-        name: "GameWorldRecordGrid",
+        name: "WorldRecordGrid",
         props: {
             gameid: String,
             categorytypeid: String,
@@ -81,7 +81,7 @@
                 var that = this;
                 this.loading = true;
 
-                axios.get('/SpeedRun/GetGameWorldRecordGridData', { params: { gameID: this.gameid, categoryTypeID: this.categorytypeid, categoryID: this.categoryid, levelID: this.levelid } })
+                axios.get('/SpeedRun/GetWorldRecordGridData', { params: { gameID: this.gameid, categoryTypeID: this.categorytypeid, categoryID: this.categoryid, levelID: this.levelid } })
                     .then(res => {
                         that.tableData = res.data;
                         that.initGrid(res.data);
